[ADD] 엔딩뷰 추가
