+++ DOMO.swiftpm/Views/Final/Final1View.swift
@@ -76,7 +76,7 @@ struct Final1View: View {
                                 ScriptButtonText(text: "REPLAY")
                                     .padding(.trailing, screenWidth * 0.02)
                             }
-                            PushView(destination: FinalTempView()) {
+                            PushView(destination: Final2View()) {
                                 ScriptButtonText(text: "NEXT")
                                     .padding(.trailing, screenWidth * 0.02)
                             }
